% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_df_br.R
\name{as_df_br}
\alias{as_df_br}
\title{Create a data frame of best responses}
\usage{
as_df_br(
  players,
  p1,
  p2,
  pars,
  par1_lim,
  par2_lim,
  cons1 = NULL,
  cons2 = NULL,
  cons_common = NULL
)
}
\arguments{
\item{players}{A character vector specifying the two players of the game.}

\item{p1}{An R function describing Player 1's payoff.}

\item{p2}{An R function describing Player 2's payoff.}

\item{pars}{A character vector of length 2, which specifies the parameters each player chooses.}

\item{par1_lim}{A numerical vector of length 2, which specifies the range of the first parameter.}

\item{par2_lim}{A numerical vector of length 2, which specifies the range of the second parameter.}

\item{cons1}{A named list of parameters contained in \code{p1} that should be treated as constants, if any.}

\item{cons2}{A named list of parameters contained in \code{p2} that should be treated as constants, if any.}

\item{cons_common}{A named list of parameters contained in p1 and p2 that should be treated as constants, if any.
If \code{cons1} and \code{cons2} are exactly same, you can specify \code{cons_common}
instead of \code{cons1} and \code{cons2}.}
}
\value{
A list containing two data frames, each of which is a player's best response to the other.
}
\description{
\code{as_df_br()} creates a list of data frames of each player's best responses in
  a given parameter space.
}
\author{
Yoshio Kamijo and Yuki Yanai <yanai.yuki@kochi-tech.ac.jp>
}
