% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br_plot.R
\name{br_plot}
\alias{br_plot}
\title{Plot of best response correspondences}
\usage{
br_plot(game)
}
\arguments{
\item{game}{A "normal_form" class object created by \code{normal_form()}.}
}
\value{
A ggplot figure of the best response correspondences.
}
\description{
\code{br_plot()} creates a plot of the best response correspondences for a 2-by-2 game.
}
\seealso{
\code{\link{normal_form}}, \code{\link[ggplot2]{ggplot}}
}
\author{
Yoshio Kamijo and Yuki Yanai <yanai.yuki@kochi-tech.ac.jp>
}
