% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_path.R
\name{draw_path}
\alias{draw_path}
\title{Color specified paths in the game tree.}
\usage{
draw_path(game, actions)
}
\arguments{
\item{game}{A game defined by \code{extensive_form()}.}

\item{actions}{A list of actions of players.}
}
\value{
A ggplot object of the game tree.
}
\description{
\code{draw_path} colors paths corresponding to the
user-specified action profiles in an extensive-form game.
}
\author{
Yoshio Kamijo and Yuki Yanai \href{mailto:yanai.yuki@kochi-tech.ac.jp}{yanai.yuki@kochi-tech.ac.jp}
}
