% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eliminate_strategy.R
\name{eliminate_strategy}
\alias{eliminate_strategy}
\title{Eliminate a strategy from a normal-form game.}
\usage{
eliminate_strategy(game, player, eliminated)
}
\arguments{
\item{game}{A normal-form game object created by \code{normal_form()}.}

\item{player}{A player one of whose strategies will be eliminated.}

\item{eliminated}{A strategy to be eliminated. Case sensitive.}
}
\description{
\code{eliminate_strategy()} eliminates a strategy from a normal-form game defined by \code{norma_form()}.
}
\author{
Yoshio Kamijo and Yuki Yanai <yanai.yuki@kochi-tech.ac.jp>
}
