% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_match.R
\name{eval_match}
\alias{eval_match}
\title{Evaluate how desirable a matching result is}
\usage{
eval_match(x, name, group = NULL, preference)
}
\arguments{
\item{x}{An object of the "matching" class obtained by \code{matching} or
\code{matching_df()}.}

\item{name}{A character string of the name of a player to be evaluated.}

\item{group}{This argument determines whether the specified player belongs to
the proposers or the proposed. If a proposing player and a proposed
player have the same name, this argument must be either "proposer" or
"proposed" in order to uniquely identify the player.  If all players have
a unique name, this can be set to \code{NULL}. Default is \code{NULL}.}

\item{preference}{A vector representing the preference of the player.
The first, second, third, ... element of the vector is the players' first,
second, third, ... choice. For example, If the player preferes A to B to
C, this argument should be \code{c("A", "B", "C")}.}
}
\value{
An integer value indicating the rank of the match given their true
preference. NA will be returned if there is no match.
}
\description{
\code{eval_match()} evaluates how desirable a matching result is
for a specified player.
}
\author{
Yoshio Kamijo and Yuki Yanai \href{mailto:yanai.yuki@kochi-tech.ac.jp}{yanai.yuki@kochi-tech.ac.jp}
}
