% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_dominant.R
\name{find_dominant}
\alias{find_dominant}
\title{Find dominant strategies in a normal-form game}
\usage{
find_dominant(game)
}
\arguments{
\item{game}{A normal-form game object created by \code{normal_form()}.}
}
\description{
\code{find_dominant()} finds each player's dominant and weakly dominant strategies if they exist.
}
\author{
Yoshio Kamijo and Yuki Yanai <yanai.yuki@kochi-tech.ac.jp>
}
