% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_dominated.R
\name{find_dominated}
\alias{find_dominated}
\title{Find dominated strategies in a normal-form game}
\usage{
find_dominated(game)
}
\arguments{
\item{game}{A normal-form game object created by \code{normal_form()}.}
}
\description{
\code{find_dominated()} finds each player's dominated
    and weakly dominated strategies if they exist.
}
\author{
Yoshio Kamijo and Yuki Yanai <yanai.yuki@kochi-tech.ac.jp>
}
