% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mixed_NE.R
\name{find_mixed_NE}
\alias{find_mixed_NE}
\title{Find a mixed-strategy Nash equilibrium}
\usage{
find_mixed_NE(game)
}
\arguments{
\item{game}{A "normal_form" class object created by \code{normal_form()}.}
}
\value{
A list of the probabilities given to each strategy that specifies the mixed-strategy Nash equilibrium.
}
\description{
\code{find_mixed_NE} finds a mixed-strategy Nash equilibrium.
}
\seealso{
\code{\link{normal_form}}
}
\author{
Yoshio Kamijo and Yuki Yanai <yanai.yuki@kochi-tech.ac.jp>
}
