% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_pure_NE.R
\name{find_pure_NE}
\alias{find_pure_NE}
\title{Find pure-strategy Nash equilibria}
\usage{
find_pure_NE(game)
}
\arguments{
\item{game}{A "normal_form" class object created by \code{normal_form()}.}
}
\value{
A character string showing the pure-strategy Nash equilibria.
}
\description{
\code{find_pure_NE()} finds pure-strategy Nash equilibria.
}
\seealso{
\code{\link{normal_form}}
}
