% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_table.R
\name{game_table}
\alias{game_table}
\title{Create a table of the game}
\usage{
game_table(game, mark_br = TRUE, cell_width = 80)
}
\arguments{
\item{game}{A "normal_form" class object created by \code{normal_form()}.}

\item{mark_br}{A logical value. If \code{TRUE}, the best response to each of the opponent's strategy is marked.
Default is \code{TRUE}.}

\item{cell_width}{A number specifying the cell width of the game matrix. The unit is pixel. Default is 80.}
}
\value{
A gt table of the normal form game.
}
\description{
\code{game_table()} creates and displays a gt table of the game.
}
\seealso{
\code{\link{normal_form}}, \code{\link[gt]{gt}}
}
