% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_payoff.R
\name{get_payoff}
\alias{get_payoff}
\title{Get payoffs corresponding to specified strategies or actions.}
\usage{
get_payoff(game, actions, cons1, cons2, cons_common)
}
\arguments{
\item{game}{A game defined by one of \code{normal_form()}, \code{seq_form()},
or \code{extensive_form()}.}

\item{actions}{A list of strategies or actions to which the payoffs
correspond. Each strategy or action must be defined in the game.}

\item{cons1}{A named list of parameters contained in \code{game$payoff$p1}
that should be treated as constants, if any.}

\item{cons2}{A named list of parameters contained in \code{game$payoff$p2}
that should be treated as constants, if any.}

\item{cons_common}{A named list of parameters contained in
\code{game$payoff$p1} and \code{game$payoff$p2} that should be treated as
constants, if any. If \code{cons1} and \code{cons2} are exactly same, you
can specify \code{cons_common} instead of specifying both \code{cons1}
and \code{cons2}.}
}
\value{
A vector of payoffs.
}
\description{
\code{get_payoff} returns the vector of payoffs corresponding to
a specified pair of strategies or a set of action profiles.
}
\author{
Yoshio Kamijo and Yuki Yanai \href{mailto:yanai.yuki@kochi-tech.ac.jp}{yanai.yuki@kochi-tech.ac.jp}
}
