% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridsearch_br.R
\name{gridsearch_br}
\alias{gridsearch_br}
\title{Search a pair of the best responses of two players}
\usage{
gridsearch_br(
  players,
  p1,
  p2,
  pars,
  par1_lim,
  par2_lim,
  cons1 = NULL,
  cons2 = NULL,
  cons_common = NULL,
  precision = 1L
)
}
\arguments{
\item{players}{A character vector specifying the two players of the game.}

\item{p1}{An R function describing Player 1's payoff.}

\item{p2}{An R function describing Player 2's payoff.}

\item{pars}{A character vector of length 2, which specifies the parameters each player chooses.}

\item{par1_lim}{A numerical vector of length 2, which specifies the range of the first parameter.}

\item{par2_lim}{A numerical vector of length 2, which specifies the range of the second parameter.}

\item{cons1}{A named list of parameters contained in \code{p1} that should be treated as constants, if any.}

\item{cons2}{A named list of parameters contained in \code{p2} that should be treated as constants, if any.}

\item{cons_common}{A named list of parameters contained in \code{p1} and \code{p2} that should be
treated as constants, if any. If \code{cons1} and \code{cons2} are exactly same, you can specify
\code{cons_common} instead of \code{cons1} and \code{cons2}.}

\item{precision}{A natural number specifying the precision of numerical approximation.
The value n approximately means that the approximation is correct up to the Nth decimal place.
The default value is 1.}
}
\value{
A data frame containing the pair of the best responses of the players.
}
\description{
\code{gridsearch_br()} numerically finds a pair of best responses of
    two players by grid search.
}
\author{
Yoshio Kamijo and Yuki Yanai <yanai.yuki@kochi-tech.ac.jp>
}
