% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_df.R
\name{matching_df}
\alias{matching_df}
\title{Implement matching}
\usage{
matching_df(
  df1,
  df2,
  df_type = NULL,
  header = NULL,
  sep = NULL,
  algorithm = "DA",
  verbose = TRUE
)
}
\arguments{
\item{df1}{A data frame or a data file containing preferences of
the proposers.}

\item{df2}{A data frame or a data file containing preferences of
the proposed.}

\item{df_type}{Type of \code{df1} and \code{df2}, which should be one of
\code{"data.frame"}, \code{"csv"}, \code{"tsv"}, or \code{"table"}.
If not specified, \code{matching_df()} tries to determine the type.}

\item{header}{A logical value showing if the data file has the header
(variable names) row. If not specified, default is \code{TRUE} for
"csv" and "tsv", and \code{FALSE} for "table".}

\item{sep}{The field separator character. Values on each line of the
file are separated by this character. If not specified,
default is \code{","} for "csv", \code{"\t"} for "tsv", and \code{""}
(white space) for "table".}

\item{algorithm}{A algorithm for matching. At the moment, \code{"DA"} is
the only available option.}

\item{verbose}{If \code{TRUE}, matching steps will be printed on screen.
Default to \code{TRUE}.}
}
\value{
A list of "matching" class containing
(1) a data frame of the matching results,
(2) a character string of the matching results, and
(3) a character string of the history of matching steps.
}
\description{
\code{matching_df()} implements matching between two groups of
individuals with preferences specified by either a pair ofdata frames or
a pair of data files (csv, tsv, or table).
}
\examples{

\dontrun{
  test7 <- matching_df(
    df1 = "foo1.csv",
    df2 = "foo2.csv")
}
}
\author{
Yoshio Kamijo and Yuki Yanai \href{mailto:yanai.yuki@kochi-tech.ac.jp}{yanai.yuki@kochi-tech.ac.jp}
}
