% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_mt1.R
\name{matching_mt1}
\alias{matching_mt1}
\title{Implement many-to-one matching}
\usage{
matching_mt1(
  df_many,
  df_one,
  capacity = NULL,
  capacity_df = NULL,
  df_type = NULL,
  header = NULL,
  sep = NULL,
  algorithm = "DA",
  verbose = TRUE
)
}
\arguments{
\item{df_many}{A data frame or a data file containing preferences of
the proposers.}

\item{df_one}{A data frame or a data file containing preferences of
the proposed.}

\item{capacity}{A value specifies how many people each of the proposed
accepts as their matches. Give a value to this parameter when the
capacity is constant among the proposed. Otherwise, specify
\code{capacity_df} instead.}

\item{capacity_df}{A data frame specifying how many people each of the
proposed accepts as their matches. The first column must be the name
of the proposed, which must match the first column of \code{df_one}.
The second column is the capacity for each of the proposed.}

\item{df_type}{Type of \code{df_many} and \code{df_one}, which should be one of
\code{"data.frame"}, \code{"csv"}, \code{"tsv"}, or \code{"table"}.
If not specified, \code{matching_df()} tries to determine the type.}

\item{header}{A logical value showing if the data file has the header
(variable names) row. If not specified, default is \code{TRUE} for
"csv" and "tsv", and \code{FALSE} for "table".}

\item{sep}{The field separator character. Values on each line of the
file are separated by this character. If not specified,
default is \code{","} for "csv", \code{"\t"} for "tsv", and \code{""}
(white space) for "table".}

\item{algorithm}{A algorithm for matching. \code{"DA"}
(\code{"Gale-Shapley"},  \code{"GS"}) or \code{"Boston"}.}

\item{verbose}{If \code{TRUE}, matching steps will be printed on screen.
Default to \code{TRUE}.}
}
\value{
A list of "matching" class containing
(1) an extended data frame of the matching results,
(2) a character string showing which algorithm was used,
(3) a character string of the matching results,
(4) a character string of the history of matching steps,
(5) a list of preferences of each group,
(6) a data frame of the matching results cleaned.
}
\description{
\code{matching_mt1()} implements many-to-one matching between
two groups of individuals with preferences specified by either a pair of
data frames or a pair of data files (csv, tsv, or table).
}
\author{
Yoshio Kamijo and Yuki Yanai \href{mailto:yanai.yuki@kochi-tech.ac.jp}{yanai.yuki@kochi-tech.ac.jp}
}
