% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_matching_data.R
\name{read_matching_data}
\alias{read_matching_data}
\title{Read data for matching}
\usage{
read_matching_data(data, df_type = NULL, header = NULL, sep = NULL)
}
\arguments{
\item{data}{A data frame or a data file containing preferences of
proposers or proposed.}

\item{df_type}{Type of \code{df1} and \code{df2}, which should be one of
\code{"data.frame"}, \code{"csv"}, \code{"tsv"}, or \code{"table"}.
If not specified, \code{matching_df()} tries to determine the type.}

\item{header}{A logical value showing if the data file has the header
(variable names) row. If not specified, default is \code{TRUE} for
"csv" and "tsv", and \code{FALSE} for "table".}

\item{sep}{The field separator character. Values on each line of the
file are separated by this character. If not specified,
default is \code{","} for "csv", \code{"\t"} for "tsv", and \code{""}
(white space) for "table".}
}
\value{
A data frame.
}
\description{
\code{read_matching_data()} reads a file to be used in
matching_df() or matching_mt1().
}
\author{
Yoshio Kamijo and Yuki Yanai \href{mailto:yanai.yuki@kochi-tech.ac.jp}{yanai.yuki@kochi-tech.ac.jp}
}
