% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrict_action.R
\name{restrict_action}
\alias{restrict_action}
\title{Restrict actions of an extensive-form game}
\usage{
restrict_action(game, action)
}
\arguments{
\item{game}{An "extensive_form" class object created by
\code{extensive_form()}.}

\item{action}{A named list of actions that can be chosen by players. The
names must be node names: "n1", "n2", etc. A player can choose any action
at the nodes that are not listed in this argument.}
}
\value{
A "restricted_game" class object in which players have to choose the
actions specified by the user.
}
\description{
\code{restrict_actions} restricts some actions of an extensive-
form game defined by \code{extensive_form()}.
}
\author{
Yoshio Kamijo and Yuki Yanai \href{mailto:yanai.yuki@kochi-tech.ac.jp}{yanai.yuki@kochi-tech.ac.jp}
}
