% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgamer.R
\docType{package}
\name{rgamer}
\alias{rgamer}
\title{rgamer: A package to help students learn Game Theory using R.}
\description{
The rgamer package provides useful functions to highlight some important aspects of
basic games. It visually displays the game and finds solutions of the games.
}
