% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_extensive.R
\name{seq_extensive}
\alias{seq_extensive}
\title{Transform a sequential_form into extensive_form}
\usage{
seq_extensive(game, ...)
}
\arguments{
\item{game}{A sequential_form object created by \code{seq_form()}. The game's
type must be "matrix".}

\item{...}{Other parameter used to define an extensive-form game.}
}
\value{
An object of "extensive_form" class.
}
\description{
\code{seq_extensive} transforms a seq_form game defined by
\code{seq_form} into an extensive_form game.
}
\examples{
g1 <- seq_form(
  players = c("Leader", "Follower"),
  s1 = c("R", "S", "P"),
  s2 = c("R", "S", "P"),
  payoffs1 = c(0, -1, 1, 1, 0, -1, -1, 1, 0),
  payoffs2 = c(0, 1, -1, -1, 0, 1, 1, -1, 0))
g1e <- seq_extensive(g1)
g1e2 <- seq_extensive(g1, direction = "right", color_palette = "Accent")
}
\seealso{
\code{\link{extensive_form}}
}
\author{
Yoshio Kamijo and Yuki Yanai \href{mailto:yanai.yuki@kochi-tech.ac.jp}{yanai.yuki@kochi-tech.ac.jp}
}
