% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_path.R
\name{show_path}
\alias{show_path}
\title{Plot a game tree with solution paths colored.}
\usage{
show_path(x, id = NULL, all = FALSE)
}
\arguments{
\item{x}{An extensive_form object defined by \code{extensive_form()} or its
solution (extensive_sol object) found by \code{solve_efg()}.}

\item{id}{An integer value vector of id numbers to specify which solutions to
show. By default, the first solution is displayed. This parameter is
ignored when \code{all = TRUE}.}

\item{all}{A logical value to determine if all solutions will be displayed.
Default is \code{FALSE}.}
}
\value{
A ggplot object of the specified game tree.
}
\description{
\code{show_path} displays the paths played under a solution.
}
\author{
Yoshio Kamijo and Yuki Yanai \href{mailto:yanai.yuki@kochi-tech.ac.jp}{yanai.yuki@kochi-tech.ac.jp}
}
