% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_EWA.R
\name{sim_EWA}
\alias{sim_EWA}
\title{Simulates an Experienced Weighted Attraction model}
\usage{
sim_EWA(
  game,
  n_periods,
  lambda = 1,
  delta = 0.5,
  rho = 0.5,
  phi = 0.5,
  A1_init = 0,
  A2_init = 0,
  N_init = 0
)
}
\arguments{
\item{game}{An object of \code{normal_form} class defined by
\code{normal_form()}.}

\item{n_periods}{A positive integer specifying how many times the game is
played within each sample.}

\item{lambda}{A positive real value representing the players' sensitivity to
attraction values of strategies. As \code{lambda} gets larger, the choice
will be dependent on attraction values more heavily.  As \code{lambda}
gets close to 0, a strategy will tend to be chosen randomly.}

\item{delta}{A real number between 0 and 1. This parameter controls how fast
attraction values of strategies that are not chosen are updated.  If
\code{delta = 0}, attraction is updated only for the strategy that is
selected at the given period (i.e., reinforcement learning is
implemented). If \code{delta = 1}, attraction is updated equally for all
strategies (i.e., belief-based learning model is applied).}

\item{rho}{A real value between 0 and 1. This parameter controls the learning
speed. \code{rho = 0} for "reinforcement" leaning and "belief" based
learning.}

\item{phi}{A real value between 0 and 1. This parameter controls how much
attraction values at the current period are constrained by the past
attraction values. If \code{phi = 0}, the past attraction values are
ignored. \code{phi = 1} for "reinforcement" leaning and "belief" based
learning.}

\item{A1_init}{An initial value of Player 1's attraction for each strategy.}

\item{A2_init}{An initial value of Player 2's attraction for each strategy.}

\item{N_init}{An initial value of N.}
}
\description{
\code{sim_EWA()} simulates plays of a normal-form
game expected by an experienced weighted attraction (EWA) model.
}
\details{
Simulate plays of a normal-form game defined by
\code{normal_form()} in a way expected by an EWA model.
}
\author{
Yoshio Kamijo and Yuki Yanai \href{mailto:yanai.yuki@kochi-tech.ac.jp}{yanai.yuki@kochi-tech.ac.jp}
}
