% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_fict.R
\name{sim_fict}
\alias{sim_fict}
\title{Simulates a fictitious play}
\usage{
sim_fict(
  game,
  n_samples = 1,
  n_periods = 50,
  lambda = 1,
  init = NULL,
  sigma = 0,
  plot_range_y = NULL,
  plot_id = 1
)
}
\arguments{
\item{game}{An object of \code{normal_form} class defined by
\code{normal_form()}.}

\item{n_samples}{A positive integer specifying the number of samples to be
simulated.}

\item{n_periods}{A positive integer specifying how many times the game is
played.}

\item{lambda}{A positive real value representing the sensitivity to
expected utilities. As \code{lambda} gets larger, a small difference in
expected utility makes a big difference in choice probability.}

\item{init}{A list of initial levels of beliefs. The length of the list must
be two. Each element must be a vector of values between 0 and 1
representing a player's belief. If \code{init = NULL}, which is default,
initial beliefs will be randomly assigned.}

\item{sigma}{A non-negative value determining the level of noise adherent to
evaluation of payoffs.}

\item{plot_range_y}{Choose the range of vertical axis for plots. Available
choices are \code{"fixed"}, \code{"full"} and \code{"free"}.
If \code{plot_range_y = "free"}, the range of y-axis depends on
simulation results.  If \code{plot_range_y = "full"}, The range
defined in \code{game} is used for each player, which can be different
between players. With \code{"fixed"}, the same y-axis is used for both
players.}

\item{plot_id}{An integer between 1 and n_samples to specify for which
sample, playing history should be displayed. If \code{plot_id = NULL},
plot_B and plot_P will be NULL.}
}
\value{
A list containing (1) a data frames of strategies chosen by each
player,  (2) a single long data frame of (1)'s data frames combined,
(3) a list of each player's belief about the opponent's behavior
(data frames), (4) a list of probability of each strategy being chosen
(data frames), and (5) three plots of simulation result.
}
\description{
\code{sim_fict()} simulates a fictitious play of a game
}
\details{
Simulate fictitious plays of a normal-form game defined by
\code{normal_form()}.
}
\author{
Yoshio Kamijo and Yuki Yanai \href{mailto:yanai.yuki@kochi-tech.ac.jp}{yanai.yuki@kochi-tech.ac.jp}
}
