% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_game.R
\name{sim_game}
\alias{sim_game}
\title{Plays a normal-form game by simulation}
\usage{
sim_game(
  game,
  n_samples,
  n_periods,
  type = "br",
  init1 = NULL,
  init2 = NULL,
  omega = 0,
  eta = 0.1,
  lambda = 1,
  cons1 = NULL,
  cons2 = NULL,
  plot_range_y = NULL
)
}
\arguments{
\item{game}{An object of \code{normal_form} class defined by
\code{normal_form()}.}

\item{n_samples}{A positive integer specifying the number of samples to be
simulated.}

\item{n_periods}{A positive integer specifying how many times the game is
played within each sample.}

\item{type}{A character string to tell what kind of simulation should be run.
The available options are \code{"br"}, \code{"sbr"}, \code{"abr"}, and
\code{"imitation"}. With \code{"br"}, each player chooses the best
response to the opponent's choice in the previous period. With
\code{"sbr"}, each player chooses the softly best response to the
opponent's choice in the previous periods, With \code{"abr"}, each
player alternately chooses the best response to the other player's
previous action. With \code{"imitation"}, each player imitates the
opponent's choice in the previous period. Players randomly choose their
strategies or the first period in each of these options.}

\item{init1}{Player 1's first strategy. If not specified, a strategy is
randomly selected from the player's strategy set.}

\item{init2}{Player 2's first strategy. If not specified, a strategy is
randomly selected from the player's strategy set.}

\item{omega}{A numeric value between 0 and 1 to control the degree of inertia
in each player's behavior. If \code{omega = 1}, each player does not
change their choices over time. If \code{omega = 0}, which is the default
value, each player does not stick to their previous choice at all.}

\item{eta}{A numeric value between 0 and 1 to control the degree of
randomness in each player's behavior. If \code{eta = 1}, each player
chooses their strategy completely at random. If \code{eta = 0}, each
player chooses the best strategy based on the opponent's behavior in the
previous period.}

\item{lambda}{A positive value controlling the weight of the best response to
the previous move of the opponent.}

\item{cons1}{A named list of parameters contained in
\code{game$payoff$payoffs1} that should be treated as constants, if any.}

\item{cons2}{A named list of parameters contained in
\code{game$payoff$payoffs2} that should be treated as constants, if any.}

\item{plot_range_y}{Choose the range of vertical axis for plots. Available
choices are \code{"fixed"}, \code{"full"} and \code{"free"}.
If \code{plot_range_y = "free"}, the range of y-axis depends on
simulation results.  If \code{plot_range_y = "full"}, The range
defined in \code{game} is used for each player, which can be different
between players. With \code{"fixed"}, the same y-axis is used for both
players.}
}
\value{
A data frame of simulation results.
}
\description{
\code{sim_game()} simulates plays expected in a normal-form
game.
}
\details{
Simulate plays expected in a normal-form game defined by
\code{normal_form()}.
}
\author{
Yoshio Kamijo and Yuki Yanai \href{mailto:yanai.yuki@kochi-tech.ac.jp}{yanai.yuki@kochi-tech.ac.jp}
}
