% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_game_br.R
\name{sim_game_br}
\alias{sim_game_br}
\title{Play a normal-form game by simulation (best response)}
\usage{
sim_game_br(
  game,
  n_periods,
  init1 = NULL,
  init2 = NULL,
  omega = 0,
  cons1 = NULL,
  cons2 = NULL
)
}
\arguments{
\item{game}{An object of \code{normal_form} class defined by
\code{normal_form()}.}

\item{n_periods}{A positive integer specifying how many times the game is
played within each sample.}

\item{init1}{Player 1's first strategy. If not specified, a strategy is
randomly selected from the player's strategy set.}

\item{init2}{Player 2's first strategy. If not specified, a strategy is
randomly selected from the player's strategy set.}

\item{omega}{A numeric value between 0 and 1 to control the degree of inertia
in each player's behavior. If \code{omega = 1}, ach player does not
change their choices over time. If \code{omega = 0}, each player does not
stick to their previous choice at all.}

\item{cons1}{A named list of parameters contained in
\code{game$payoff$payoffs1} that should be treated as constants, if any.}

\item{cons2}{A named list of parameters contained in
\code{game$payoff$payoffs2} that should be treated as constants, if any.}
}
\value{
data.frame containing the history of the game played.
}
\description{
\code{sim_game()} simulates plays expected in a normal-form
game.
}
\details{
Simulate plays expected in a normal-form game defined by
\code{normal_form()} when each player chooses the best response to the
other player's previous move.
}
\author{
Yoshio Kamijo and Yuki Yanai \href{mailto:yanai.yuki@kochi-tech.ac.jp}{yanai.yuki@kochi-tech.ac.jp}
}
