% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_game_abr.R
\name{simu_game_abr}
\alias{simu_game_abr}
\title{Plays a normal-form game by simulation (alternate best response)}
\usage{
simu_game_abr(game, n_periods, rho = 0.2, cons1 = NULL, cons2 = NULL)
}
\arguments{
\item{game}{An object of \code{normal_form} class defined by \code{normal_form()}.}

\item{n_periods}{A positive integer specifying how many times the game is played within each sample.}

\item{rho}{A numeric value in [0, 1] to control the degree of inertia in each player's behavior. If \code{rho = 1},
each player does not change their choices over time. If \code{rho = 0}, each player does not stick to their
previous choice at all.}

\item{cons1}{A named list of parameters contained in \code{game$payoff$p1} that should be treated as constants, if any.}

\item{cons2}{A named list of parameters contained in \code{game$payoff$p2} that should be treated as constants, if any.}
}
\value{
data.frame containing the history of the game played.
}
\description{
\code{simu_game_abr()} simulates plays expected in a normal-form game.
}
\details{
Simulate plays expected in a normal-form game defined by \code{normal_form()} when
  each player alternately moves and each chose the best response to the opponent's previous
  move..
}
\author{
Yoshio Kamijo and Yuki Yanai <yanai.yuki@kochi-tech.ac.jp>
}
