% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_efg.R
\name{solve_efg}
\alias{solve_efg}
\title{Find solutions of an extensive-form game.}
\usage{
solve_efg(game, concept = "backward", quietly = FALSE)
}
\arguments{
\item{game}{An "extensive_form" class object created by}

\item{concept}{Solution concept to be used. It must be one of
\code{"backward"} (backward induction) or \code{"spe"} (subgame perfect
equilibrium).}

\item{quietly}{A logical value. If \code{TRUE}, the solution(s) will not be
displayed on screen. Default is \code{FALSE}.}
}
\value{
A list containing (1) \code{sols}: a list of solutions, if any, (2)
\code{n_sols}: the number of solutions found, and (3) \code{trees}: a
list of game trees each of which shows which paths were played under
each solution.
}
\description{
\code{solve_efg} finds solutions of an extensive-form game
defined by \code{extensive_form()}.
}
\examples{
g1 <- extensive_form(
  players = list("Kamijo",
                 rep("Yanai", 2),
                 rep(NA, 4)),
  actions <- list(c("U", "D"),
                  c("U'", "D'"), c("U''", "D''")),
  payoffs = list(Kamijo = c(0, 2, 1, 3),
                 Yanai  = c(0, 1, 2, 1)))
s1 <- solve_efg(g1)

g2 <- extensive_form(
  players = list("f",
                 c("m", "m"),
                 rep(NA, 4)),
  actions = list(c("ballet", "baseball"),
                 c("ballet", "baseball"), c("ballet", "baseball")),
  payoffs = list(f = c(2, 0, 0, 1),
                 m = c(1, 0, 0, 2)),
  show_node_id = FALSE)
s2 <- solve_efg(g2, quietly = TRUE)
}
\author{
Yoshio Kamijo and Yuki Yanai \href{mailto:yanai.yuki@kochi-tech.ac.jp}{yanai.yuki@kochi-tech.ac.jp}
}
