% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_nfg_char.R
\name{solve_nfg_char}
\alias{solve_nfg_char}
\title{Find Nash equilibria of a normal-form game with  payoffs defined by character strings}
\usage{
solve_nfg_char(game, delta = 0.1, plot = TRUE, quietly = FALSE)
}
\arguments{
\item{game}{A "normal_form" class object created by \code{normal_form()}.}

\item{delta}{A numerical value specifying the grid size to draw the figure of best response correspondences.
The default value is 0.1. The smaller the value is, the smoother the correspondence curves are.}

\item{plot}{A logical value to determine whether the figure of the best response correspondences
will be displayed. Default is \code{TRUE}.}

\item{quietly}{A logical value to determine if the equilibrium will be kept in the returned list
without being printed on screen. Default is \code{FALSE}.}
}
\value{
A list containing the pair of the best response correspondence (NE)  and the plot of best
    response correspondences.
}
\description{
\code{solve_nfg_char()} finds the pair of the best responses when payoff
    functions are provided as character strings.
}
\seealso{
\code{\link{normal_form}}
}
