% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_seq.R
\name{solve_seq}
\alias{solve_seq}
\title{Find Nash equilibria of a sequential-form game}
\usage{
solve_seq(
  game,
  show_table = TRUE,
  mark_br = FALSE,
  cons1 = NULL,
  cons2 = NULL,
  cons_common = NULL,
  precision = 1L,
  quietly = FALSE
)
}
\arguments{
\item{game}{A "sequential_form" class object created by \code{seq_form()}.}

\item{show_table}{A logical value. If \code{TRUE}, the table of the game will
be displayed. Default is \code{TRUE}.}

\item{mark_br}{A logical value. If \code{TRUE}, the follower's best response
to each of the leader's strategy is marked. Default is \code{FALSE}.}

\item{cons1}{A named list of parameters contained in \code{game$payoff$p1} that should be treated as constants, if any.}

\item{cons2}{A named list of parameters contained in \code{game$payoff$p2} that should be treated as constants, if any.}

\item{cons_common}{A named list of parameters contained in \code{game$payoff$p1}
and \code{game$payoff$p2} that should be treated as constants, if any.
If \code{cons1} and \code{cons2} are exactly same, you can specify
\code{cons_common} instead of specifying both \code{cons1} and \code{cons2}.}

\item{precision}{A natural number specifying the precision of numerical approximation.
The value n approximately means that the approximation is correct up to the Nth decimal place.
The default value is 1.}

\item{quietly}{A logical value that determines whether the equilibrium will
be kept in the returned list without being printed on screen. Default is
\code{FALSE}.}
}
\value{
A list containing Nash equilibrium outcomes (if any), and
the game table (if available).
}
\description{
\code{solve_seq()} finds Nash equilibrium outcomes of a
sequential-form game. This is a wrapper function of
\code{\link{solve_seq_matrix}}, \code{\link{solve_seq_char}}, and
\code{\link{solve_seq_fcn}}.
}
\author{
Yoshio Kamijo and Yuki Yanai \href{mailto:yanai.yuki@kochi-tech.ac.jp}{yanai.yuki@kochi-tech.ac.jp}
}
