% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_seq_char.R
\name{solve_seq_char}
\alias{solve_seq_char}
\title{Find subgame perfect equilibria outcome of a sequential-form game
(extensive form game) with  payoffs defined by character strings}
\usage{
solve_seq_char(game, precision = 1L, quietly = FALSE)
}
\arguments{
\item{game}{A "sequential_form" class object created by \code{seq_form()}.}

\item{precision}{A natural number specifying the precision of numerical approximation.
The value n approximately means that the approximation is correct up to the Nth decimal place.
The default value is 1.}

\item{quietly}{A logical value to determine if the equilibrium will be kept in the returned list
without being printed on screen. Default is \code{FALSE}.}
}
\value{
A list containing the Nash equilibrium (NE) outcomes
and the plot of best response correspondences.
}
\description{
\code{solve_seq_char()} finds the subgame perfect equilibrium
outcomes when payoff functions are provided as character strings.
}
\seealso{
\code{\link{seq_form}}
}
\author{
Yoshio Kamijo and Yuki Yanai \href{mailto:yanai.yuki@kochi-tech.ac.jp}{yanai.yuki@kochi-tech.ac.jp}
}
