% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_seq_matrix.R
\name{solve_seq_matrix}
\alias{solve_seq_matrix}
\title{Find subgame perfect equilibria of a sequential-form game.}
\usage{
solve_seq_matrix(
  game,
  show_table = TRUE,
  mark_br = FALSE,
  cell_width = NULL,
  quietly = FALSE
)
}
\arguments{
\item{game}{A "sequential_form" class object created by \code{seq_form()}.}

\item{show_table}{A logical value. If \code{TRUE}, the table of the game will
be displayed. Default is \code{TRUE}.}

\item{mark_br}{A logical value. If \code{TRUE}, the follower's best response
to each of the leader's strategy is marked. Default is \code{FALSE}.}

\item{cell_width}{A number specifying the cell width of the game matrix. The
unit is pixel.}

\item{quietly}{A logical value that determines whether the equilibrium will
be kept in the returned list without being printed on screen. Default is
\code{FALSE}.}
}
\value{
A list containing subgame perfect equilibrium (SPE) outcome and
the gt table of the game.
}
\description{
\code{solve_seq_matrix()} finds subgame perfect equilibria of
a sequential-form (an extensive-form) game with discrete-choice strategies.
}
\seealso{
\code{\link{seq_form}}
}
\author{
Yoshio Kamijo and Yuki Yanai \href{mailto:yanai.yuki@kochi-tech.ac.jp}{yanai.yuki@kochi-tech.ac.jp}
}
