% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgames.R
\name{subgames}
\alias{subgames}
\title{Find subgames of an extensive-form game}
\usage{
subgames(game, quietly = FALSE)
}
\arguments{
\item{game}{An extensive-form game defined by \code{extensive_form()}.}

\item{quietly}{If TRUE, a message telling how many subgames are found will
not be displaed.}
}
\value{
A list of subgames.
}
\description{
\code{subgames} finds subgames contained in an extensive-form
game.
}
\details{
This function finds subgames of an extensive-form game defined by
\code{extensive_form()} and returns the list of subgames. The node IDs of
each subgame displayed with the game tree are taken from the original
extensive-form game. In other words, the node ID of a subgame in the tree
does not necessarily begins with 1.
}
\author{
Yoshio Kamijo and Yuki Yanai \href{mailto:yanai.yuki@kochi-tech.ac.jp}{yanai.yuki@kochi-tech.ac.jp}
}
