% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_matrix.R
\name{to_matrix}
\alias{to_matrix}
\title{Transform an extensive-form game into a normal-form game.}
\usage{
to_matrix(game)
}
\arguments{
\item{game}{A two-person extensive-form game defined by
\code{extensive_form()}.}
}
\value{
A normal_form class object.
}
\description{
\code{to_matrix} transoforms a two-person extensive-form game
into a normal-form game
a specified pair of strategies or a set of action profiles.
}
\author{
Yoshio Kamijo and Yuki Yanai \href{mailto:yanai.yuki@kochi-tech.ac.jp}{yanai.yuki@kochi-tech.ac.jp}
}
