% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_position.R
\name{tree_position}
\alias{tree_position}
\title{Define the node positions of a game tree}
\usage{
tree_position(players, n_node, n_choice)
}
\arguments{
\item{players}{A list of players.  Each element of the list must correspond to each game node, except for
the terminal (payoff) nodes.}

\item{n_node}{A vector of the number of nodes at each sequence, including the terminal (payoff) nodes.}

\item{n_choice}{A list of the number of choices at each node. An element of the list must
correspond to each sequence (including the terminal node as the final sequence). Each element must
be a numeric vector whose length equals the number of nodes at the specific sequence.}
}
\value{
A data frame containing the positions of each node on x and y axes..
}
\description{
\code{tree_position} determines the position of nodes in a game tree.
}
\details{
Create a data frame necessary to draw a game tree
}
\author{
Yoshio Kamijo and Yuki Yanai <yanai.yuki@kochi-tech.ac.jp>
}
