% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dom.R
\name{dom}
\alias{dom}
\title{Find dominated and dominant strategies in a normal-form game}
\usage{
dom(game, type = "dominated")
}
\arguments{
\item{game}{A normal-form game object created by \code{normal_form()}.}

\item{type}{Specify which to be found, "dominated" or "dominant."
Default is "dominated."}
}
\description{
\code{dom()} finds each player's (weakly) dominated
    or (weakly) dominant strategies
}
\author{
Yoshio Kamijo and Yuki Yanai <yanai.yuki@kochi-tech.ac.jp>
}
