% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensive_form.R
\name{extensive_form}
\alias{extensive_form}
\title{Define and solve a extensive-form (or strategic-form) game}
\usage{
extensive_form(
  players = NULL,
  n_node,
  n_choice,
  strategy,
  payoff,
  quietly = FALSE,
  show_tree = TRUE,
  mark_path = FALSE,
  direction = "right"
)
}
\arguments{
\item{players}{A list of players. Each element of the list must be a character string
or a vector of character strings at a specific sequence.}

\item{n_node}{A numeric vector of the number of nodes at a specific sequence.}

\item{n_choice}{A list of the number of choices at each node. Each element of the list must be
a numeric vector of the numbers of choices at a specific sequence. You must assign \code{0}
for the terminal (payoff) nodes.}

\item{strategy}{A list of strategies. Each element of the list must be a vector of character strings
that corresponds to a specific node.}

\item{payoff}{A named list of payoffs. Each element of the list must be a numeric vector of
payoffs for a player. The names of the elements must match the names of the players
specified  by \code{players}.}

\item{quietly}{A logical value. If \code{TRUE}, the subgame perfect equilibrium will not be displayed
on screen. Default is \code{FALSE}.}

\item{show_tree}{A logical value. If \code{TRUE}, the game tree will be displayed. Default is \code{TRUE}}

\item{mark_path}{A logical value. If \code{TRUE}, The paths played in the equilibrium will be
marked (with color and bold lines). Default is \code{FALSE}.}

\item{direction}{The direction to which a game tree grows. The value must be one of
\code{"right"}, \code{"up"}, and \code{"down"}. Default is \code{"right"}.}
}
\value{
An object of "extensive_form" class, which defines an extensive-form (or sequential) game.
}
\description{
\code{extensive_form} defines an extensive-form game
}
\examples{
g1 <- extensive_form(
  players = list("Kamijo",
                 rep("Yanai", 2)),
  n_node = c(1, 2, 4),
  n_choice = list(2,
                  c(2, 2),
                  rep(0, 4)),
  strategy <- list(c("U", "D"),
                   c("U'", "D'"), c("U''", "D''")),
  payoff = list(Kamijo = c(0, 2, 1, 3),
                Yanai  = c(0, 1, 2, 1)))

g2 <- extensive_form(
  players = list("f", c("m", "m")),
 n_node = c(1, 2, 4),
 n_choice = list(2,
                 c(2, 2),
                 rep(0, 4)),
  strategy = list(c("ballet", "baseball"),
                 c("ballet", "baseball"), c("ballet", "baseball")),
 payoff = list(f = c(2, 0, 0, 1),
               m = c(1, 0, 0, 2)),
 mark_path = TRUE,
 quietly = TRUE
)

g3 <- extensive_form(
  players = list("p1", rep("p2", 3)),
  n_node = c(1, 3, 6),
  n_choice = list(3,
                  rep(2, 3),
                  rep(0, 6)),
  strategy = list(c("C", "D", "E"),
                  c("F", "G"), c("H", "I"), c("J", "K")),
  payoff = list(p1 = c(3, 1, 1, 2, 2, 1),
                p2 = c(0, 0, 1, 1, 2, 3)),
  mark_path = TRUE,
  direction = "down")

g4 <- extensive_form(
  players = list("child",
                 "parent"),
   n_node = c(1, 2, 2),
   n_choice = list(2,
                   c(0, 2),
                   c(0, 0)),
   strategy = list(c("give up", "keep asking"),
                   c("leave",   "buy")),
   payoff   = list(child  = c(0, -10, 10),
                   parent = c(5, -10,  2)),
   mark_path = TRUE)
}
\author{
Yoshio Kamijo and Yuki Yanai <yanai.yuki@kochi-tech.ac.jp>
}
