% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_stable.R
\name{is_stable}
\alias{is_stable}
\title{Check if a matching is stable}
\usage{
is_stable(x, verbose = FALSE)
}
\arguments{
\item{x}{An object of the "matching" class obtained by \code{matching} or
\code{matching_df()}.}

\item{verbose}{If \code{TRUE}, the result and blocking pairs will be printed
on screen. Default is \code{FALSE}.}
}
\value{
A list containing (1) a logical value showing if a matching is
stable and (2) a data frame containing the information of blocking pairs,
if any.
}
\description{
\code{is_stable()} checks if a matching result is stable.
}
\author{
Yoshio Kamijo and Yuki Yanai \href{mailto:yanai.yuki@kochi-tech.ac.jp}{yanai.yuki@kochi-tech.ac.jp}
}
