% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.matching.R
\name{print.matching}
\alias{print.matching}
\title{Custom print method for "matching" class}
\usage{
\method{print}{matching}(x, type = "results")
}
\arguments{
\item{x}{The output of \code{matching()}.}

\item{type}{What to print. The available options are "results", "history",
and "both" (both results and history). Default to \code{"results"}.}
}
\description{
Custom print method for "matching" class created by
\code{matching()}.
}
\author{
Yoshio Kamijo and Yuki Yanai \href{mailto:yanai.yuki@kochi-tech.ac.jp}{yanai.yuki@kochi-tech.ac.jp}
}
