% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_fict.R
\name{sim_fict}
\alias{sim_fict}
\title{Simulates a fictitious play}
\usage{
sim_fict(game, n_periods = 50, lambda = 1, init = NULL, sigma = 0)
}
\arguments{
\item{game}{An object of \code{normal_form} class defined by
\code{normal_form()}.}

\item{n_periods}{A positive integer specifying how many times the game is
played.}

\item{lambda}{A positive real value representing the sensitivity to
expected utilities. As \code{lambda} gets larger, a small difference in
expected utility makes a big difference in choice probability.}

\item{init}{A list of initial levels of beliefs. The length of the list must
be two. Each element must be a vector of values between 0 and 1
representing a player's belief. If \code{init = NULL}, which is default,
initial beliefs will be randomly assigned.}

\item{sigma}{A non-negative value determining the level of noise adherent to
evaluation of payoffs.}
}
\value{
A list containing (1) a data frames of strategies chosen by each
player, (2) a list of each player's belief about the opponent's behavior
(data frames), (3) a list of probability of each strategy being chosen
(data frames), and (4) a plot of cumulative ratio of each strategy
chosen, (5) a plot of beliefs, and (6) a plot of choice probabilities.
}
\description{
\code{sim_fict()} simulates a fictitous play of a game
}
\details{
Simulate fictitious plays of a normal-form game defined by
\code{normal_form()}.
}
\author{
Yoshio Kamijo and Yuki Yanai \href{mailto:yanai.yuki@kochi-tech.ac.jp}{yanai.yuki@kochi-tech.ac.jp}
}
