% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_nfg_char.R
\name{solve_nfg_char}
\alias{solve_nfg_char}
\title{Find Nash equilibria of a normal-form game.}
\usage{
solve_nfg_char(
  game,
  plot = TRUE,
  mark_NE = FALSE,
  quietly = FALSE,
  color_palette = "Set1"
)
}
\arguments{
\item{game}{A "normal_form" class object created by \code{normal_form()}.}

\item{plot}{A logical value to determine whether the figure of the best
response correspondences will be displayed. Default is \code{TRUE}.}

\item{mark_NE}{A logical value to control if the NE (if any) will be marked
in the best response plot, which will be displayed (only displayed when
\code{plot = TRUE}). Default is \code{FALSE}.}

\item{quietly}{A logical value to determine if the equilibrium will be kept
in the returned list without being printed on screen. Default is
\code{FALSE}.}

\item{color_palette}{A color palette to be used. Default is \code{"Set1"}.}
}
\value{
A list containing the pair of the best responses of two players (NE)
and the plot of best response correspondences.
}
\description{
\code{solve_nfg_char()} finds the pair of the best responses
of a normal-form game defined with payoff functions as character
strings by \code{normal_form()}.
}
\seealso{
\code{\link{normal_form}}
}
\author{
Yoshio Kamijo and Yuki Yanai \href{mailto:yanai.yuki@kochi-tech.ac.jp}{yanai.yuki@kochi-tech.ac.jp}
}
