% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_nfg_matrix.R
\name{solve_nfg_matrix}
\alias{solve_nfg_matrix}
\title{Find Nash equilibria of a normal-form game.}
\usage{
solve_nfg_matrix(
  game,
  mixed = FALSE,
  show_table = TRUE,
  mark_br = TRUE,
  quietly = FALSE,
  color_palette = "Set1"
)
}
\arguments{
\item{game}{A "normal_form" class object created by \code{normal_form()}.}

\item{mixed}{A logical value. If \code{TRUE}, mixed-strategy NE will be
searched. Default is \code{FALSE}.}

\item{show_table}{A logical value. If \code{TRUE}, the payoff matrix  of the
game will be displayed. Default is \code{TRUE}.}

\item{mark_br}{A logical value. If \code{TRUE}, the best response to each of
the opponent's strategy is marked. Default is \code{TRUE}.}

\item{quietly}{A logical value that determines whether the equilibrium will
be kept in the returned list without being printed on screen. Default is
\code{FALSE}.}

\item{color_palette}{A color palette to be used. Default is \code{"Set1"}.}
}
\value{
A list containing pure strategy Nash equilibria (NE), mixed strategy
NE, a table of the payoff matrix, the plot of best response
correspondences, and the list of mixed strategy NE of the subsets of
strategies. Each element will be \code{NULL} if not available.
}
\description{
\code{solve_nfg_matrix()} finds Nash equilibria of a normal-form
(strategic-form) game with discrete-choice strategies.
}
\seealso{
\code{\link{normal_form}}
}
\author{
Yoshio Kamijo and Yuki Yanai \href{mailto:yanai.yuki@kochi-tech.ac.jp}{yanai.yuki@kochi-tech.ac.jp}
}
